+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sit-onyx
 
+## 1.0.0-alpha.103
+
+### Patch Changes
+
+- 09ce727: fix(OnyxCheckbox): align checkbox label padding according to density, stabilize check all border
+
+## 1.0.0-alpha.102
+
+### Minor Changes
+
+- 3b9fbf4: feat: add `OnyxTextarea` component
+
 ## 1.0.0-alpha.101
 
 ### Patch Changes